+++ src/index.ts
@@ -1,7 +1,7 @@
 export { Builder } from './builder'
 export { ASTERISK, DISTINCT, ALL } from './singletoneConstants'
 export { LogicalOperator } from './steps'
-export { e, o, f } from './functions'
+export { e, o, f, $ } from './functions'
 export {
   ComparisonOperator,
   ArithmeticOperator,